Extract RepositoryBuilder
