+++ src/Konmaripo.Web.Tests.Unit/Services/GitHubServiceTests.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using FluentAssertions;
 using Konmaripo.Web.Services;
+using Konmaripo.Web.Tests.Unit.Helpers;
 using Moq;
 using Octokit;
 using Xunit;
@@ -60,44 +61,5 @@ namespace Konmaripo.Web.Tests.Unit.Services
             }
         }
 
-        public class RepositoryBuilder
-        {
-            private string _repositoryName = "";
-            const string DummyString = "";
-            const long DummyLong = 0;
-            const int DummyInt = 0; 
-            private readonly DateTimeOffset DummyDateTimeOffset;
-
-            public RepositoryBuilder()
-            {
-                DummyDateTimeOffset = DateTimeOffset.Now;
-            }
-
-            public RepositoryBuilder WithName(string repoName)
-            {
-                _repositoryName = repoName;
-
-                return this;
-            }
-
-            public Repository Build()
-            {
-                var dummyInternalRepository = new Repository(1);
-                var licenseMetadata = new LicenseMetadata(DummyString, DummyString, DummyString, DummyString, DummyString, false);
-                var repositoryPermissions = new RepositoryPermissions(false, false, false);
-                var owner = new User();
-
-                var repo = new Repository(DummyString, DummyString, DummyString, DummyString, DummyString, DummyString,
-                    DummyString, DummyLong, DummyString, owner, _repositoryName, DummyString, false, DummyString,
-                    DummyString, DummyString, false, false, DummyInt, DummyInt, DummyString, DummyInt, DummyDateTimeOffset,
-                    DummyDateTimeOffset, DummyDateTimeOffset, repositoryPermissions,
-                    dummyInternalRepository, dummyInternalRepository,
-                    licenseMetadata,
-                    false, false, false, false, 0, 0, false, false, false, false);
-
-                return repo;
-            }
-
-        }
     }
 }
